# src/go-server-server-generated/go/model_apis.go
/*
 * zhihu_caffe
 *
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * API version: 0.1.4
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package swagger

type Apis struct {

	UsersUrl string `json:"users_url,omitempty"`

	UserUrl string `json:"user_url,omitempty"`

	ArticlesUrl string `json:"articles_url,omitempty"`

	ArticleUrl string `json:"article_url,omitempty"`

	RepliesUrl string `json:"replies_url,omitempty"`

	ReplyUrl string `json:"reply_url,omitempty"`

	TagsUrl string `json:"tags_url,omitempty"`

	TagUrl string `json:"tag_url,omitempty"`
}