# src/go-server-server-generated/go/model_article.go
/*
 * zhihu_caffe
 *
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * API version: 0.1.4
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package swagger

type Article struct {

	Id int64 `json:"id,omitempty"`

	Title string `json:"title,omitempty"`

	ReadNum int64 `json:"readNum,omitempty"`

	LikeNum int64 `json:"likeNum,omitempty"`

	Content string `json:"content,omitempty"`

	Replies []Reply `json:"replies,omitempty"`

	Tags []Tag `json:"tags,omitempty"`

	Url string `json:"url,omitempty"`
}